pyproject.toml

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepbrowse"
version = "0.1.0"
description = "Deep browse and research framework."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Umer Haider", email = "umerhaider@gmail.com"}
]

# Core dependencies
dependencies = [
    "aiohttp>=3.11.14",
    "beautifulsoup4>=4.13.3",
    "brotli>=1.1.0",
    "browser-use>=0.1.40",
    "httpx>=0.28.1",
    "langchain-community>=0.3.20",
    "langchain[openai]>=0.3.22",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
